fix: bug fixes + missing features + Render deployment fixes
- BUG-5: Order modal defaults to chart symbol instead of EUR_USD
- BUG-6: Reduce polling (TopBar 15s->60s, Trading 3s->5s)
- BUG-7: Prevent broker-auto-connect firing on every navigation (useRef guard)
- BUG-9: Remove duplicate Ctrl+K handler from ChatSidebar
- BUG-11: Download Backup uses authenticated fetch+blob (JWT support)
- Fix [object Object] rendering in StrategyEditor error display
- Fix system strategies deleted by _remove_incompatible_strategies (add is_system guard)
- Fix CORS: always include localhost origins, handle trailing slashes

Missing features:
- Data row preview: expandable table showing first 10 candle rows
- Live bid/ask/spread display in New Order modal
- ML Model dropdown in Edit Agent modal

+++ backend/app/models/datasource.py
@@ -18,7 +18,7 @@ class DataSource(Base):
     date_from = Column(String(30), default="")
     date_to = Column(String(30), default="")
     columns = Column(String(500), default="")         # Comma-separated column names
-    file_size_mb = Column(Integer, default=0)
+    file_size_mb = Column(Float, default=0)
     source_type = Column(String(20), default="upload")  # upload, broker
     broker_name = Column(String(20), default="")        # mt5, oanda, coinbase, tradovate
     # Instrument profile
